+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project are documented in this file.
 - Documentation for extending `bmcmanager` with new features.
 - Reduced Docker image size.
 - `bmcmanager server list` command, to list available servers.
+- `timeout` option for NetBox DCIM.
 
 ### Changed